+++ src/app/@shared/models/formData.model.ts
@@ -1,0 +1,9 @@
+export class FormDataModel {
+  formID: string;
+  formData: any[];
+
+  constructor(options: { formID: string; formData?: any[] }) {
+    this.formID = options.formID;
+    this.formData = options.formData;
+  }
+}